+++ lib/tokencost.dart
@@ -1,4 +1,6 @@
-/// A Very Good Project created by Very Good CLI.
+/// Costs dictionary and utility tool for counting tokens
+// library tokencost;
 library tokencost;
 
+export 'src/constants.dart';
 export 'src/tokencost.dart';